requirements.txt
streamlit
python-decouple
langchain
langchain-openai
langchain-chroma
langchain-community
chromadb
tiktoken
pypdf
